QueryString check for BadRequests done in controllers

+++ ProjectVishnu/ServerApp/App/Controllers/ObrasController.cs
@@ -19,6 +19,7 @@ namespace ProjectVishnu.Controllers
         [HttpGet]
         [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<ObraOutputModel>))]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public IActionResult List([FromQuery(Name = "mercado")] string? mercado)
         {
             IEnumerable<Obra> obraList;
@@ -26,10 +27,14 @@ namespace ProjectVishnu.Controllers
             {
                 obraList = _obrasService.ListAlphabetically();
             }
-            else
+            else if(!Request.QueryString.HasValue)
             {
                 obraList = _obrasService.ListByMarket(mercado);
 
+            }
+            else
+            {
+                return BadRequest();
             }
             return obraList == null ? NotFound() : Ok(obraList.Select(obra => obra.toObraOutputModel()));
         }
@@ -37,8 +42,10 @@ namespace ProjectVishnu.Controllers
         [HttpGet("{codigoInterno}")]
         [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ObraOutputModel))]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public IActionResult Get(string codigoInterno)
         {
+            if (Request.QueryString.HasValue) return BadRequest();
             Obra result = _obrasService.Get(codigoInterno);
             return result == null ? NotFound() : Ok(result.toObraOutputModel());
             
@@ -49,6 +56,7 @@ namespace ProjectVishnu.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public IActionResult Create([FromBody] ObraInputModel obraInput) 
         {
+            if (Request.QueryString.HasValue) return BadRequest();
             string codigoInterno = _obrasService.Create(obraInput);
             var actionName = nameof(ObrasController.Get);
             var routeValues = new
@@ -64,6 +72,7 @@ namespace ProjectVishnu.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public IActionResult Edit(string codigoInterno, [FromBody] ObraInputModel obraInput) 
         {
+            if (Request.QueryString.HasValue) return BadRequest();
             string result = _obrasService.Update(codigoInterno,obraInput);
             return result == null ? NotFound() : Ok(result);
         }
@@ -73,6 +82,7 @@ namespace ProjectVishnu.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public IActionResult Delete(string codigoInterno)
         {
+            if (Request.QueryString.HasValue) return BadRequest();
             string result = _obrasService.Delete(codigoInterno);
             return result == null ? NotFound() : Ok(result);
         }
